strrchr

+++ ft_memcpy.c
@@ -6,14 +6,14 @@
 /*   By: nhuang <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/09 13:28:02 by nhuang            #+#    #+#             */
-/*   Updated: 2023/03/09 14:53:53 by nhuang           ###   ########.fr       */
+/*   Updated: 2023/03/19 20:23:40 by nhuang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-#include"libft.h"
+
 
 void	ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (count < n)
